Added locked default value (false) to store method and modified the update and destroy methods to check the locked attribute before applying changes. If the Instance is locked then a 403 error is thrown.
Relates to #356

+++ app/Http/Controllers/V2/InstanceController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\V2\CreateInstanceRequest;
 use App\Http\Requests\V2\UpdateInstanceRequest;
 use App\Models\V2\Instance;
 use App\Resources\V2\InstanceResource;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use UKFast\DB\Ditto\QueryTransformer;
 
@@ -52,7 +53,8 @@ class InstanceController extends BaseController
      */
     public function store(CreateInstanceRequest $request)
     {
-        $instance = new Instance($request->only(['network_id', 'name']));
+        $instance = new Instance($request->only(['vpc_id', 'name', 'vpc_id', 'locked']));
+        $instance->locked = false;
         $instance->save();
         $instance->refresh();
         return $this->responseIdMeta($request, $instance->getKey(), 201);
@@ -66,7 +68,11 @@ class InstanceController extends BaseController
     public function update(UpdateInstanceRequest $request, string $instanceId)
     {
         $instance = Instance::forUser(app('request')->user)->findOrFail($instanceId);
-        $instance->fill($request->only(['vpc_id', 'name']));
+        if ((!$request->has('locked') || $request->get('locked') !== false) &&
+            (bool) $instance->locked === true) {
+            return $this->isLocked();
+        }
+        $instance->fill($request->only(['vpc_id', 'name', 'vpc_id', 'locked']));
         $instance->save();
         return $this->responseIdMeta($request, $instance->getKey(), 200);
     }
@@ -79,7 +85,24 @@ class InstanceController extends BaseController
     public function destroy(Request $request, string $instanceId)
     {
         $instance = Instance::forUser($request->user)->findOrFail($instanceId);
+        if ((bool) $instance->locked === true) {
+            return $this->isLocked();
+        }
         $instance->delete();
         return response()->json([], 204);
+    }
+
+    /**
+     * @return \Illuminate\Http\JsonResponse
+     */
+    private function isLocked(): JsonResponse
+    {
+        return JsonResponse::create([
+            'errors' => [
+                'title'  => 'Forbidden',
+                'detail' => 'The specified instance is locked',
+                'status' => 403,
+            ]
+        ], 403);
     }
 }
